Rework component selection reason to support composite explanations
This commit changes how `ComponentSelectionReason` works, so that we can save a list of selection
reasons, instead of a single one. This should let us give richer explananations about why a component
was selected. In particular, a component may be selected because another one was rejected, or it
can be selected by a rule **and** by conflict resolution. This last case was handled in an adhoc
manner, it's now a regular case.

Signed-off-by: Cedric Champeau <cedric@gradle.com>